Remove unnecessary method "updateTask"
The only way it is usefull is if using the console to update multiple
fields but it cannot be accessed there because webpack wraps all the
code in an IIFE. Im not used to this limitation which is why I came up
with that method. I also wanted a reason to use destructuring syntax.