Define a funciton that retrieves the time left
Im referring to the time left to complete the task.

Remove direct access to "dueDate" property. Directly assigning a new
value to the date property of a task object was causing problems; by
updating only the instance property and not the original variable, I
could not get "timeLeft" to work. The original value was always empty.

Note to self: I don't fully understand that situation at this moment,
but the code is working and I would like to move on.