Define a method that moves a category
As with the other methods that move an objects place within an array,
"moveCategories" invokes a method that prints objects to the console.

Information from "printCategories" should be expandable. I dont know
why I didn't make the output the same at the other print methods.

The message variable should not be in quotes. Fix that as well.