Save and retrieve data from sessionStorage
Assign timestamps to categories, projects, and tasks in order to store
and retrieve their data from sessionStorage individually. All setter
methods should update the corresponding object in sessionStorage.

More information was required in the printTasks output to solve a
problem where information was being duplicated. Rename the function to
to not confuse it with the other 'printTask' functions.

Modify start-up demonstration code and move it to to-do.js file because
I find myself moving back and forth with these changes, specifically,
manually giving each of these objects a fake timestamp because they're
all made at the same time. Timestamps serve as ids and must be unique.
Eventually, I thought of an easy way to go about assigning timestamps
but I still dont want it to be in the main.js module.

Adjust the code as necessary.