Adjust the appearance for larger screen sizes
An attribute was missing on the html meta tag that would allow chrome to
make use of the responsive utilities and it appears that things were not
being scaled properly so the root font size has to be readjusted. Border
sizes also need readjustment.

Hide the gear Icon. It does not look good.

Add a link to the left panel that redirects to project github page.